Add VM interface break stuff up and refactor DetailVM
